# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/products": {
      "get": {
        "operationId": "ProductsController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Products"
        ]
      },
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/admin/all": {
      "get": {
        "operationId": "ProductsController_all",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "operationId": "ProductsController_one",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Products"
        ]
      }
    },
    "/api/products/{id}/variants/{variantId}/availability": {
      "get": {
        "operationId": "ProductsController_availability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação concluída",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Availability"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida"
          },
          "401": {
            "description": "Sessão ausente ou inválida"
          },
          "403": {
            "description": "Sem permissão"
          },
          "404": {
            "description": "Recurso não encontrado"
          },
          "409": {
            "description": "Estado ou estoque incompatível"
          },
          "429": {
            "description": "Limite de requisições"
          }
        },
        "security": [],
        "tags": [
          "Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Apollo Rigor API",
    "description": "Vendas e locações avulsas. Valores em centavos; datas YYYY-MM-DD. Checkout exclusivamente simulado. Auth intermediado pelo NestJS/Supabase.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "VariantDto": {
        "type": "object",
        "properties": {
          "size": {
            "type": "string",
            "minLength": 1,
            "maxLength": 20,
            "pattern": "\\S"
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000
          }
        },
        "required": [
          "size",
          "quantity"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "pattern": "\\S"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "collection": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "fabric": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "color": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "line": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "photoUrl": {
            "type": "string",
            "format": "uri"
          },
          "rentalPriceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "salePriceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "variants": {
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            }
          }
        },
        "required": [
          "name",
          "category",
          "rentalPriceCents",
          "salePriceCents",
          "variants"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120,
            "pattern": "\\S"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "collection": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "fabric": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "color": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "line": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "photoUrl": {
            "type": "string",
            "format": "uri"
          },
          "rentalPriceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "salePriceCents": {
            "type": "number",
            "minimum": 0,
            "maximum": 100000000
          },
          "variants": {
            "minItems": 1,
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VariantDto"
            }
          }
        }
      },
...
      "Variant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "size": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "collection": {
            "type": "string"
          },
          "fabric": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "line": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          },
          "rentalPriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "salePriceCents": {
            "type": "integer",
            "minimum": 0,
            "example": 10000
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variant"
            }
          }
        }
      },
...
      "Availability": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer"
          },
          "reserved": {
            "type": "integer"
          },
          "available": {
            "type": "integer"
          },
          "hasConflict": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}